refactor(shared): extract share link sanitizer
